Update Pjoin packet

+++ src/NosCore.Controllers/GroupPacketController.cs
@@ -46,7 +46,7 @@ namespace NosCore.Controllers
 
                 case GroupRequestType.Requested:
                 case GroupRequestType.Invited:
-                    if (pjoinPacket.CharacterId == 0 || targetSession == null || (long)pjoinPacket.CharacterId == Session.Character.CharacterId)
+                    if (pjoinPacket.CharacterId == 0 || targetSession == null || pjoinPacket.CharacterId == Session.Character.CharacterId)
                     {
                         return;
                     }
@@ -69,7 +69,7 @@ namespace NosCore.Controllers
                         return;
                     }
 
-                    if (Session.Character.IsRelatedToCharacter((long)pjoinPacket.CharacterId, CharacterRelationType.Blocked))
+                    if (Session.Character.IsRelatedToCharacter(pjoinPacket.CharacterId, CharacterRelationType.Blocked))
                     {
                         Session.SendPacket(new InfoPacket
                         {
@@ -87,7 +87,7 @@ namespace NosCore.Controllers
                         return;
                     }
 
-                    Session.Character.GroupRequestCharacterIds.Add((long)pjoinPacket.CharacterId);
+                    Session.Character.GroupRequestCharacterIds.Add(pjoinPacket.CharacterId);
 
                     if (Session.Character.Group == null || Session.Character.Group.Type == GroupType.Group)
                     {
@@ -96,8 +96,8 @@ namespace NosCore.Controllers
                             targetSession.SendPacket(new DlgPacket
                             {
                                 Question = Language.Instance.GetMessageFromKey(LanguageKey.INVITED_YOU_GROUP, targetSession.Account.Language),
-                                YesPacket = new PjoinPacket { CharacterId = (ulong)Session.Character.CharacterId, RequestType = GroupRequestType.Accepted },
-                                NoPacket = new PjoinPacket { CharacterId = (ulong)Session.Character.CharacterId, RequestType = GroupRequestType.Declined }
+                                YesPacket = new PjoinPacket { CharacterId = Session.Character.CharacterId, RequestType = GroupRequestType.Accepted },
+                                NoPacket = new PjoinPacket { CharacterId = Session.Character.CharacterId, RequestType = GroupRequestType.Declined }
                             });
                         }
                     }
